Added Jack-In option for REPL startup
This commit adds the ability to start a REPL and automatically connect
to it.  The command to start the REPL is configurable.  The default
command assumes that sbcl is being used with quicklisp.  It also sets
the path of the REPL to the root of the project on startup.  This should
make it easier to use the starting path for environment names in CLPM
and to make things less confusing in general.  This setup should be
compatible with starting the repl in CLPM and Roswell without
forcing the user into using either.