Added automatic REPL config and startup (#58)
This commit adds the ability to download Swank if it isn't already
downloaded, add it to the ASDF path, start Swank, and attach the REPL to
it.  The command to start the REPL is configurable.  The default command
uses SBCL, but it can be overridden in the config.  The mechanism for
adding Swank should work with all lisps that use ASDF 3.x.

It sets the path of the REPL to the root of the project on startup.
This should make it easier to use the starting path for environment
names in CLPM and to make things less confusing in general.  This setup
should be compatible with starting the repl in CLPM and Roswell without
forcing the user into using either.

Ultimately, this should make it so that VSCode users can get a REPL
via the editor instead of having to manually start it up first.

It adds two dependencies because I had a hell of a time with the limited
functionality of the built-in node HTTP libraries and zlib libraries.  I
apologize. I just couldn't do what I wanted to do without rewriting two
fairly hefty libs.