REPL startup failure indicator
In cases where the REPL has not yet attached and the output to the Swank
Trace output has gone longer than 3 seconds without receiving an update,
a warning is displayed and the Swank Trace output window is opened.

By default, nothing other than the warning happens.  This is done to
keep from disrupting scenarios where the REPL takes a long time to
start, but does eventually succeed.  Three seconds is a complete
ass-pull.  If it needs to be longer or shorter based on real info, it
will need to be changed.

If an attempt to attach a REPL is run against a Swank server that hasn't
connected, it reports back that the REPL is still trying to connect.  If
closing REPL is called, it will get it back into a state where
connecting can be tried again.

closes https://github.com/nobody-famous/alive/issues/61